+++ packages/DropdownMenu/src/components/LinkItem/LinkItem.js
@@ -40,11 +40,8 @@ const LinkItem = props => {
     linkItemProps.suffixIcon = null;
   }
 
-  return (
-    <Button.Link css={linkItemStyles} {...linkItemProps}>
-      {children}
-    </Button.Link>
-  );
+  // must fix applying of css styles
+  return <Button.Link {...linkItemProps}>{children}</Button.Link>;
 };
 
 LinkItem.displayName = "DropdownMenu.LinkItem";