+++ packages/DropdownMenu/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+- Updated Fixed LinkItem styles when focused (now using Button.Link)[@tristanjasper](https://github.com/tristanjasper).
+
 ## [2.0.40] - 2020-09-02
 
 ### Added